monster commit. moves everything to check folder and loads correct files
