+++ intrigue-issues.gemspec
@@ -7,7 +7,12 @@ Gem::Specification.new do |s|
   s.description = "Intrigue Core Issues"
   s.authors     = ["jcran"]
   s.email       = 'jcran@intrigue.io'
-  s.files       = Dir.glob("lib/issues*.rb").concat Dir.glob("lib/checks/*.rb").concat ["lib/intrigue-issues.rb"]
+  s.files       = Dir.glob("lib/issues*.rb").concat 
+                  Dir.glob("lib/checks/*.rb").concat 
+                  Dir.glob("lib/tasks/base.rb").concat 
+                  Dir.glob("lib/issues/base.rb").concat
+                  Dir.glob("lib/task_factory.rb").concat
+                  Dir.glob("lib/issue_factory.rb").concat ["lib/intrigue-issues.rb"]
   s.require_paths = ['./lib']
   s.homepage    = 'http://github.com/intrigueio/intrigueio/intrigue-core'
   s.license     = 'BSD-3-Clause'
