+++ lib/all_base.rb
@@ -1,0 +1,5 @@
+require_relative 'task_factory'
+require_relative 'issue_factory'
+require_relative 'tasks/base'
+require_relative 'issues/base'
+require_relative 'checks/base'
