+++ lib/checks/base.rb
